ethereum utils: added more functionality
